change from oracle to mssql

+++ orca/update_scripts/quote.py
@@ -53,7 +53,7 @@ class QuoteUpdater(UpdaterBase):
 
         for dname in quote_sql.dnames:
             key = {'dname': dname, 'date': date}
-            self.db.quote.update(key, {'$set': {'dvalue': df[dname].dropna().to_dict()}}, upsert=True)
+            self.db.quote.update(key, {'$set': {'dvalue': df[dname].dropna().astype(float).to_dict()}}, upsert=True)
         logger.info('UPSERT documents for %d sids into (c: [%s]) of (d: [%s]) on %s',
                 len(df), self.db.quote.name, self.db.name, date)
